+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Fixed
 - Fixed DNS outage causing problems for DNS records
+- Fixed string values to escape TF characters in resource discovery
 
 ## 3.76.0 (May 19, 2020)